requirement.txt
Flask==1.1.1
flask-apidoc==1.2.0
Flask-HTTPAuth==3.3.0
requests==2.22.0
gunicorn==20.0.0
flask-restful==0.3.7
flask-script==2.0.6
bert-serving-client==1.10.0
tensorflow==1.15.0
scikit-learn==0.22.2.post1
keras==2.1.4
numpy==1.18.2
jieba==0.42.1
gensim==3.8.1
pyyaml==5.3.1
winney-redis==0.1.3
redis==3.5.1